Added Settings (Env + KeyBinds)
