+++ assets/templates/about.tmpl
@@ -40,6 +40,9 @@
                         <div class="input-group spacing-top">
                             <span class="input-group-addon detail-size">Version</span>
                         </div>
+                        <div class="input-group spacing-top">
+                            <span class="input-group-addon detail-size">Storage</span>
+                        </div>
                     </div>
                     <div class="column">
                         <div class="input-group spacing-top">
@@ -47,6 +50,9 @@
                         </div>
                         <div class="input-group spacing-top">
                             <span class="input-group-addon">{{ .nbversion }}</span>
+                        </div>
+                        <div class="input-group spacing-top">
+                            <span class="input-group-addon">{{ .storage }}</span>
                         </div>
                     </div>
                 </div>
