+++ assets/templates/console-error.tmpl
@@ -9,7 +9,6 @@
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <style type="text/css">
     html, body { width: 100%; height: 100%; margin: 0 }
-    .button-edit { width: 75px ! important; margin-top: 9px; margin-right: 9px; }
   </style>
 </head>
   <div style="margin-top: 15px; height: 100%;">
