+++ assets/templates/error.tmpl
@@ -1,10 +1,10 @@
-{{ $root := .root }}{{ $error := .error }}<!DOCTYPE html>
+{{ $error := .error }}{{ $llen := len .location }}<!DOCTYPE html>
 <html>
 <head>
     <meta charset="utf-8">
     <title>Rizin Notebook</title>
-    <link rel="stylesheet" href="{{ $root }}static/spectre-icons.min.css">
-    <link rel="stylesheet" href="{{ $root }}static/spectre.min.css">
+    <link rel="stylesheet" href="{{ .root }}static/spectre-icons.min.css">
+    <link rel="stylesheet" href="{{ .root }}static/spectre.min.css">
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <style type="text/css">
         .spacing-top { margin: 3px 0px 0px 0px; }
@@ -16,11 +16,11 @@
     <section class="container grid-960 mt-10">
         <header class="navbar">
             <section class="navbar-section">
-                <img src="{{ $root }}static/logo.png" width="30" height="30" />&nbsp;
-                <a href="{{ $root }}" class="navbar-brand mr-10">Rizin Notebook</a>
+                <img src="{{ .root }}static/logo.png" width="30" height="30" />&nbsp;
+                <a href="{{ .root }}" class="navbar-brand mr-10">Rizin Notebook</a>
             </section>
             <section class="navbar-section">
-                <a href="{{ $root }}" class="btn btn-primary">Go Back</a>
+                <a href="{{ if lt $llen 1 }}{{ .root }}{{ else }}{{ .location }}{{ end }}" class="btn btn-primary">Go Back</a>
             </section>
         </header>
         <section class="container">
