+++ assets/templates/index.tmpl
@@ -23,6 +23,8 @@
                 <a href="{{ $root }}" class="navbar-brand mr-10">Rizin Notebook</a>
             </section>
             <section class="navbar-section">
+                <a href="{{ $root }}settings" class="btn">Settings</a>
+                &nbsp;
                 <a href="{{ $root }}about" class="btn">About</a>
                 &nbsp;
                 <a href="{{ $root }}new" class="btn btn-primary">New Page</a>
